+++ src/pages/PostsManagerPage.tsx
@@ -23,7 +23,7 @@ import { Contents } from "../widgets/ui/Contents"
 import { ContentFilter } from "../widgets/ui/ContentFilter"
 import { PostFilter } from "../features/post/ui/PostFilter"
 import { Pagination } from "../widgets/ui/Pagination"
-import { PostTable } from "../features/post/ui/PostTable"
+import { PostTable } from "../entities/post/ui/PostTable"
 import { PostAddDialog } from "../features/post/ui/PostAddDialog"
 import { PostUpdateDialog } from "../features/post/ui/PostUpdateDialog"
 import { Comment, NewComment } from "../entities/comment/model/types"
